refactor: cleanup deprecated API leftover
