+++ extensions/control-plane/api/management-api/contract-negotiation-api/src/main/java/org/eclipse/edc/connector/api/management/contractnegotiation/ContractNegotiationApiExtension.java
@@ -79,7 +79,6 @@ public class ContractNegotiationApiExtension implements ServiceExtension {
 
         var monitor = context.getMonitor();
 
-        webService.registerResource(config.getContextAlias(), new ContractNegotiationApiController(monitor, service, transformerRegistry));
-        webService.registerResource(config.getContextAlias(), new ContractNegotiationNewApiController(service, transformerRegistry, jsonLd, monitor));
+        webService.registerResource(config.getContextAlias(), new ContractNegotiationApiController(service, transformerRegistry, jsonLd, monitor));
     }
 }
